migrate configs into shared module
